convert back to CommonJS
